requirements.txt
dnspython==2.4.0
Requests==2.31.0
rich==13.4.2
typer==0.9.0
